+++ duvida/__init__.py
@@ -4,15 +4,15 @@ appname = "duvida"
 __version__ = version(appname)
 __author__ = "Eachan Johnson"
 
-from .stateless.hessians import (
+from .hessians import (
     bekas,
     exact_diagonal,
     get_approximators, 
     rough_finite_difference,
     squared_jacobian
 )
-from .stateless.hvp import hvp
-from .stateless.information import (
+from .hvp import hvp
+from .information import (
     doubtscore,
     fisher_score,
     fisher_information_diagonal,
@@ -20,4 +20,3 @@ from .stateless.information import (
     parameter_hessian_diagonal,
     information_sensitivity
 )
-from .stateless import numpy
