+++ duvida/hessians.py
@@ -5,9 +5,9 @@ from typing import Any, Callable, Dict, Optional, Tuple, Union
 from functools import partial
 
 from .hvp import hvp
-from .numpy import numpy as dnp
-from .typing import Approximator, Array
+from .types import Approximator, Array
 from .utils import grad, random_normal, vmap, get_eps
+from .numpy import numpy as dnp
 
 _EPS = get_eps() ** .5
 _DEFAULT_APPROXIMATOR = 'exact_diagonal'
@@ -85,7 +85,7 @@ def squared_jacobian(
     Examples
     --------
     >>> from duvida.stateless.utils import hessian
-    >>> import duvida.stateless.numpy as dnp 
+    >>> import duvida.numpy as dnp  
     >>> f = lambda x: dnp.sum(x ** 3. + x ** 2. + 4.)
     >>> a = dnp.array([1., 2.])
     >>> f(a)
@@ -134,7 +134,7 @@ def exact_diagonal(
     Examples
     --------
     >>> from duvida.stateless.utils import hessian
-    >>> import duvida.stateless.numpy as dnp 
+    >>> import duvida.numpy as dnp  
     >>> f = lambda x: dnp.sum(x ** 3. + x ** 2. + 4.)
     >>> a = dnp.array([1., 2.])
     >>> exact_diagonal(f)(a) == dnp.diag(hessian(f)(a))
@@ -206,7 +206,7 @@ def bekas(
     Examples
     --------
     >>> from duvida.stateless.utils import hessian
-    >>> import duvida.stateless.numpy as dnp 
+    >>> import duvida.numpy as dnp  
     >>> f = lambda x: dnp.sum(x ** 3. + x ** 2. + 4.)
     >>> a = dnp.array([1., 2.])
     >>> f(a)
@@ -285,10 +285,8 @@ def rough_finite_difference(
 
     Examples
     --------
-    >>> from duvida.stateless.config import config
-    >>> config.set_backend("jax", precision="double")
     >>> from duvida.stateless.utils import hessian
-    >>> import duvida.stateless.numpy as dnp 
+    >>> import duvida.numpy as dnp 
     >>> f = lambda x: dnp.sum(x ** 3. + x ** 2. + 4.)
     >>> a = dnp.array([1., 2.])
     >>> f(a)
