+++ duvida/numpy/__init__.py
