+++ duvida/utils.py
@@ -6,7 +6,7 @@ from functools import partial
 from carabiner import print_err
 
 from .config import config
-from .typing import Array, ArrayLike
+from .types import Array, ArrayLike
 
 __backend__ = config.backend
